Extract Logging + Provider libraries, rename ProviderDatabase to Provider.Database with subfolders, restructure tests, apply ReSharper cleanup

+++ tests/Unit/EventLogExpert.Eventing.Tests/Interop/NativeMethodsEvtTests.cs
@@ -234,7 +234,7 @@ public sealed class NativeMethodsEvtTests
     public void ConvertVariant_WhenHexInt32_ShouldReturnInt32()
     {
         // Arrange
-        int expectedValue = unchecked((int)0x1234ABCD);
+        int expectedValue = unchecked(0x1234ABCD);
         var variant = CreateVariant(EvtVariantType.HexInt32, expectedValue);
 
         // Act
